Bug Fix and GUI Testing
Completely fixed bugs. Currently in testing for custom GUI elements, such as circular buttons

+++ src/assets/SpeedBoost.java
@@ -23,7 +23,7 @@ public class SpeedBoost implements Purchasable, Serializable {
 	
 	private final DecimalFormat currency = (DecimalFormat) NumberFormat.getCurrencyInstance();
 
-	public SpeedBoost(int value, long length, double cost, String name) {
+	public SpeedBoost(double value, long length, double cost, String name) {
 		this.value = value;
 		this.length = length;
 		this.cost = cost;
